VendingMachine_Capstone
For our first big design project, my partner and I were tasked with designing a vending machine application in the command line using Java. We designed a vending machine class to intake our vendingmachine.csv file and implemented methods() for prompting the user for making purchases.  The class design can be cleaned up a little bit more but as it stands the vending machine should be able to take any vendingfile.csm so long as it is formatted appropriately and incorporate that into our design without having to change anything aside from the input file in the main class and VendingMachine class.